Move EC functionality to a new file
